requirements.txt
pandas
transformers
jupyter
ipykernel
numpy
torch
pyabsa
NewsSentiment
python-dotenv
seaborn
matplotlib
flair
autorootcwd
chromadb
spacy-sentiws
libretranslate
xgboost
fuzzywuzzy
tensorflow